requirements.txt
asgiref==3.8.1
certifi==2025.7.9
cffi==1.17.1
charset-normalizer==3.4.2
cloudinary==1.44.1
cryptography==45.0.6
dj-rest-auth==7.0.1
Django==5.2.3
django-allauth==65.11.0
django-cloudinary-storage==0.3.0
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
git-filter-repo==2.47.0
idna==3.10
pillow==11.3.0
psycopg2-binary==2.9.10
pycparser==2.22
PyJWT==2.9.0
python-decouple==3.8
python-dotenv==1.1.1
requests==2.32.4
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.5.0
